GRHB-631: * refactor, split course navigation
