+++ mobile/src/common/enums/navigation/uam-screen-name.enum.ts
@@ -1,5 +1,5 @@
 enum UAMScreenName {
-  UAM = 'uam',
+  UAM = 'UAM',
   UAM_GROUPS_CREATE = 'Create Group',
   UAM_GROUPS_EDIT = 'Edit Group',
 }
