+++ mobile/src/components/my-courses/my-courses.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import { MyCourses as MyCoursesScreen } from '~/navigation/my-courses/my-courses-stack.navigation';
+import { MyCourses as MyCoursesScreen } from '~/navigation/my-courses/my-courses.navigation';
 
 const MyCourses: FC = () => {
   return <MyCoursesScreen />;
