+++ mobile/src/navigation/chat/chat.navigation.tsx
@@ -1,9 +1,9 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import React, { FC } from 'react';
+import React, { FC, useEffect } from 'react';
 
 import { ChatScreenName } from '~/common/enums/enums';
 import { ChatNavigationParamList } from '~/common/types/types';
-import { useAppNavigate, useCallback, useFocusEffect } from '~/hooks/hooks';
+import { useAppNavigate } from '~/hooks/hooks';
 
 import { NAVIGATION_ITEMS, SCREEN_OPTIONS } from './common/constants/constants';
 
@@ -11,13 +11,11 @@ const Chat: FC = () => {
   const NativeStack = createNativeStackNavigator<ChatNavigationParamList>();
   const navigation = useAppNavigate();
 
-  useFocusEffect(
-    useCallback(() => {
-      navigation.setOptions({
-        headerShown: true,
-      });
-    }, []),
-  );
+  useEffect(() => {
+    navigation.setOptions({
+      headerShown: true,
+    });
+  }, []);
 
   return (
     <NativeStack.Navigator screenOptions={SCREEN_OPTIONS}>
