+++ mobile/src/navigation/courses/common/constants/constants.ts
@@ -3,7 +3,7 @@ import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
 import { CoursesScreenName } from '~/common/enums/enums';
 import { AddCourse } from '~/components/courses/components/components';
 import { Courses } from '~/components/courses/courses';
-import { Course } from '~/navigation/course/course-stack.navigation';
+import { ConfigureCourse } from '~/navigation/configure-course/configure-course.navigation';
 import { CoursesNavigationItem } from '~/navigation/courses/common/types/types';
 
 const SCREEN_OPTIONS: NativeStackNavigationOptions = {
@@ -26,7 +26,7 @@ const NAVIGATION_ITEMS: CoursesNavigationItem[] = [
   },
   {
     name: CoursesScreenName.COURSE,
-    component: Course,
+    component: ConfigureCourse,
     permissions: [],
     isAuthRequired: false,
   },
