+++ mobile/src/navigation/my-courses-results/common/constants/constants.ts
@@ -1,0 +1,24 @@
+import { MaterialTopTabNavigationOptions } from '@react-navigation/material-top-tabs';
+
+import { AppColor, AppFontFamily } from '~/common/enums/enums';
+
+const SCREEN_OPTIONS: MaterialTopTabNavigationOptions = {
+  tabBarLabelStyle: {
+    fontFamily: AppFontFamily.INTER_600,
+    fontSize: 14,
+    lineHeight: 16,
+    letterSpacing: 0.5,
+    textTransform: 'none',
+  },
+  tabBarStyle: {
+    backgroundColor: AppColor.BACKGROUND.GRAY_300,
+  },
+  tabBarIndicatorStyle: {
+    backgroundColor: AppColor.BRAND.BLUE_100,
+  },
+  swipeEnabled: false,
+  tabBarActiveTintColor: AppColor.TEXT.GRAY_100,
+  tabBarInactiveTintColor: AppColor.TEXT.GRAY_200,
+};
+
+export { SCREEN_OPTIONS };
